fixed likedislike + verify user for delete and modify fonction + add images in gitignore

+++ backend/controllers/sauce.js
@@ -12,7 +12,6 @@ exports.createSauce = (req, res, next) => {
         usersLiked: [],
         usersDisliked: []
    });
-  console.log(sauce);
    sauce.save()
    .then(() =>  res.status(201).json({message: 'Objet enregistré !'}))
    .catch(error =>  res.status(400).json({message: "Problème d'enregistrement de la sauce créé !"}))
@@ -24,16 +23,20 @@ exports.modifySauce = (req, res, next) => {
     imageUrl: `${req.protocol}://${req.get('host')}/images/${req.file.filename}`
   } : { ...req.body };
 
+if(req.body.userId == sauceObject.userId) {
+  console.log(sauceObject.userId);
    Sauce.updateOne({_id: req.params.id}, { ...sauceObject, _id: req.params.id})
       .then(()=> res.status(200).json({message : 'Sauce modifiée !'}))
       .catch(error => res.status(401).json({message : "Problème de mise à jour des données de votre sauce !" }));
+  }
 };
 
 exports.deleteSauce =  (req, res, next)=>{
-
+if(req.body.userId == Sauce.userId) {
    Sauce.deleteOne({_id: req.params.id})
             .then(() => { res.status(200).json({message: 'Objet supprimé !'})})
             .catch( error => res.status(401).json({ message : "Problème pour récupérer la sauce à suppprimer !" }));
+}
 };
 
 exports.getOneSauce = (req, res, next) => {
@@ -51,13 +54,13 @@ exports.getAllSauces = (req, res, next)=> {
 exports.likeDislike = (req, res, next) => {
     if (req.body.like === 1) {
         Sauce.updateOne({_id: req.params.id}, {
-          likes: req.body.like++,
+          $inc: {likes: +1},
           usersLiked: req.body.userId })
             .then(sauce => res.status(200).json({message: 'Like ajouté !'}))
             .catch(error => res.status(400).json({message: "Problème d'ajout du like !"}))
     } else if (req.body.like === -1) {
         Sauce.updateOne({ _id: req.params.id }, {
-          dislikes: (req.body.like++) * -1,
+          $inc: {dislikes: +1},
           usersDisliked: req.body.userId })
             .then(sauce => res.status(200).json({message: 'Dislike ajouté !'}))
             .catch(error => res.status(400).json({message: "Problème d'ajout du dislike !"}))
@@ -66,14 +69,15 @@ exports.likeDislike = (req, res, next) => {
         Sauce.findOne({ _id: req.params.id })
             .then(sauce => {
                 if (sauce.usersLiked.includes(req.body.userId)) {
-                    Sauce.updateOne({ _id: req.params.id },{
-                    usersLiked: req.body.userId,
-                    likes: -1 })
+                    Sauce.updateOne({ _id: req.params.id },
+                      { $pull: {
+                    usersLiked: req.body.userId},
+                    $inc: {likes: -1 }})
                         .then(sauce => { res.status(200).json({message: 'Like supprimé !'}) })
                         .catch(error => res.status(400).json({message: "Problème de suppression du like !"}))
                 } else if (sauce.usersDisliked.includes(req.body.userId)) {
-                    Sauce.updateOne({ _id: req.params.id }, { usersDisliked: req.body.userId,
-                      dislikes: -1 })
+                    Sauce.updateOne({ _id: req.params.id },{$pull: { usersDisliked: req.body.userId},
+                      $inc: {dislikes: -1 }})
                         .then(sauce => { res.status(200).json({message: 'Dislike supprimé !'}) })
                         .catch(error => res.status(400).json({message: "Problème de suppression du dislike !"}))
                 }
